--- lib/pages/get_started/src/get_started.dart
@@ -1,1 +1,0 @@
-export '../src/widgets/app_title_text.dart';
