--- lib/pages/log_in/src/log_in.dart
@@ -1,5 +1,0 @@
-export '../src/widgets//password_text_form.dart';
-export '../src/widgets/email_text_form.dart';
-export '../src/widgets/forget_password_text_button.dart';
-export '../src/widgets/reset_email_text_form.dart';
-export '../src/widgets/reset_password_back_button.dart';
