--- lib/pages/log_in/src/widgets/reset_password_back_button.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '/utils/constants/app_colors.dart';
-
-class ResetPasswordBackButton extends StatelessWidget {
-  const ResetPasswordBackButton({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.centerLeft,
-      child: IconButton(
-        onPressed: () {
-          Navigator.of(context).pop();
-        },
-        icon: const Icon(
-          Icons.close,
-          size: 32,
-          color: AppColors.baseDark,
-        ),
-      ),
-    );
-  }
-}
