+++ lib/pages/sign_up/sign_up_page.dart
@@ -1,18 +1,26 @@
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 import '../../features/auth/sign_up.dart';
 import '../../gen/assets.gen.dart';
 import '../../utils/async_value_error_dialog.dart';
+import '../../utils/constants/app_colors.dart';
 import '../../utils/constants/measure.dart';
+import '../../utils/constants/string.dart';
+import '../../utils/dialog.dart';
+import '../../utils/exceptions/app_exception.dart';
 import '../../utils/loading.dart';
 import '../../utils/scaffold_messenger_service.dart';
+import '../../utils/text_form_styles.dart';
+import '../../utils/text_styles.dart';
 import '../../widgets/gradation_background.dart';
 import '../../widgets/rounded_button.dart';
+import '../../widgets/text_form_header.dart';
 import '../../widgets/white_app_bar.dart';
 import '../home/home_page.dart';
-import 'src/sign_up.dart';
 
 class SignUpPage extends HookConsumerWidget {
   const SignUpPage({super.key});
@@ -136,6 +144,216 @@ class SignUpPage extends HookConsumerWidget {
             ),
           ),
           if (ref.watch(overlayLoadingProvider)) const OverlayLoadingWidget(),
+        ],
+      ),
+    );
+  }
+}
+
+class UserNameTextForm extends StatelessWidget {
+  const UserNameTextForm({
+    required this.controller,
+    super.key,
+  });
+
+  final TextEditingController controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: Measure.p_h32,
+      child: Column(
+        children: [
+          const TextFormHeader(title: 'ユーザ名'),
+          Measure.g_4,
+          TextFormField(
+            maxLength: 20,
+            controller: controller,
+            decoration: AppTextFormStyles.onGeneral(
+              iconData: Icons.account_circle,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+
+class EmailTextForm extends StatelessWidget {
+  const EmailTextForm({
+    required this.controller,
+    super.key,
+  });
+
+  final TextEditingController controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: Measure.p_h32,
+      child: Column(
+        children: [
+          const TextFormHeader(title: 'メールアドレス'),
+          Measure.g_4,
+          TextFormField(
+            controller: controller,
+            decoration: AppTextFormStyles.onGeneral(
+              iconData: Icons.mail,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+class PasswordTextForm extends StatelessWidget {
+  const PasswordTextForm({
+    required this.controller,
+    required this.isObscure,
+    super.key,
+  });
+
+  final TextEditingController controller;
+  final ValueNotifier<bool> isObscure;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: Measure.p_h32,
+      child: Column(
+        children: [
+          const TextFormHeader(title: 'パスワード'),
+          Measure.g_4,
+          TextFormField(
+            obscureText: isObscure.value,
+            controller: controller,
+            decoration: AppTextFormStyles.onPassword(
+              isObscure: isObscure,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+class TermsAndPrivacyPolicyText extends HookConsumerWidget {
+  const TermsAndPrivacyPolicyText({
+    required this.isCheckTerms,
+    super.key,
+  });
+
+  final ValueNotifier<bool> isCheckTerms;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Padding(
+      padding: Measure.p_h48,
+      child: Row(
+        children: [
+          Checkbox(
+            activeColor: AppColors.secondary,
+            checkColor: AppColors.baseWhite,
+            value: isCheckTerms.value,
+            onChanged: (value) => isCheckTerms.value = value!,
+          ),
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              RichText(
+                text: TextSpan(
+                  children: <TextSpan>[
+                    TextSpan(
+                      text: '利用規約 ',
+                      recognizer: TapGestureRecognizer()
+                        ..onTap = () async {
+                          ref
+                              .watch(
+                                overlayLoadingProvider.notifier,
+                              )
+                              .update((state) => true);
+                          try {
+                            if (!await launchUrl(
+                              Uri.parse(termsLink),
+                              mode: LaunchMode.externalApplication,
+                            )) {
+                              const exception = AppException(
+                                message: 'Could not launch $termsLink',
+                              );
+                              throw exception;
+                            }
+                          } on AppException catch (e) {
+                            await showAlertDialog(
+                              context: context,
+                              title: 'Error',
+                              content: e.message,
+                              defaultActionText: 'OK',
+                            );
+                          } finally {
+                            ref
+                                .watch(
+                                  overlayLoadingProvider.notifier,
+                                )
+                                .update((state) => false);
+                          }
+                        },
+                      style: TextStyles.p2(
+                        color: AppColors.secondary,
+                      ),
+                    ),
+                    TextSpan(
+                      text: 'と',
+                      style: TextStyles.p2(),
+                    ),
+                    TextSpan(
+                      text: ' プライバシー・ポリシー ',
+                      recognizer: TapGestureRecognizer()
+                        ..onTap = () async {
+                          ref
+                              .watch(
+                                overlayLoadingProvider.notifier,
+                              )
+                              .update((state) => true);
+                          try {
+                            if (!await launchUrl(
+                              Uri.parse(privacyPolicyLink),
+                              mode: LaunchMode.externalApplication,
+                            )) {
+                              const exception = AppException(
+                                message: '''
+Could not launch $privacyPolicyLink .
+''',
+                              );
+                              throw exception;
+                            }
+                          } on AppException catch (e) {
+                            await showAlertDialog(
+                              context: context,
+                              title: 'Error',
+                              content: e.message,
+                              defaultActionText: 'OK',
+                            );
+                          } finally {
+                            ref
+                                .watch(
+                                  overlayLoadingProvider.notifier,
+                                )
+                                .update((state) => false);
+                          }
+                        },
+                      style: TextStyles.p2(
+                        color: AppColors.secondary,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              Text(
+                'に同意する。',
+                style: TextStyles.p2(),
+              ),
+            ],
+          ),
         ],
       ),
     );
